requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
chromadb
google-genai
anthropic
python-dotenv
rich
